Introduce BasicArticleEditor.
Renamed ManuscriptEditor to DefaultEditorEditor
