--- src/article/api/ManuscriptEditorAPI.js
@@ -1,9 +1,0 @@
-import ArticleAPI from './ArticleAPI'
-
-/**
- * An implementation of the otherwise generic ArticleAPI,
- * providing methods for setting selections after structural changes.
- */
-export default class ManuscriptEditorAPI extends ArticleAPI {
-
-}
