+++ src/article/components/index.js
@@ -1,10 +1,12 @@
 export { default as AbstractComponent } from './AbstractComponent'
 export { default as AddSupplementaryFileWorkflow } from './AddSupplementaryFileWorkflow'
 export { default as AuthorsListComponent } from './AuthorsListComponent'
+export { default as BasicArticleEditor } from './BasicArticleEditor'
 export { default as BlockFormulaComponent } from './BlockFormulaComponent'
 export { default as BlockQuoteComponent } from './BlockQuoteComponent'
 export { default as BoldComponent } from './BoldComponent'
 export { default as BreakComponent } from './BreakComponent'
+export { default as DefaultArticleEditor } from './DefaultArticleEditor'
 export { default as DefaultNodeComponent } from './DefaultNodeComponent'
 export { default as DOIInputComponent } from './DOIInputComponent'
 export { default as DownloadSupplementaryFileTool } from './DownloadSupplementaryFileTool'
@@ -33,7 +35,6 @@ export { default as LicenseEditor } from './LicenseEditor'
 export { default as ListComponent } from './ListComponent'
 export { default as ListItemComponent } from './ListItemComponent'
 export { default as ManuscriptComponent } from './ManuscriptComponent'
-export { default as ManuscriptEditor } from './ManuscriptEditor'
 export { default as ManuscriptSection } from './ManuscriptSection'
 export { default as ManuscriptTOC } from './ManuscriptTOC'
 export { default as MetadataFieldComponent } from './MetadataFieldComponent'
